Fix a bug in LogEntry reading code, pointed out by Jan, that prevents files with non-zero servo values from being read.

+++ src/org/openaltimeter/data/LogEntry.java
@@ -135,7 +135,7 @@ public class LogEntry {
 			temperature = Double.parseDouble(splitLine[5]);
 			battery = Double.parseDouble(splitLine[8]);	
 			if (splitLine.length >= 12) 
-				servo = Byte.parseByte(splitLine[11]);
+				servo = Integer.parseInt(splitLine[11]);
 			else
 				servo = 0;
 		} else {
